+++ src/algorithms/linear_search.c
@@ -1,6 +1,3 @@
-#include <stdlib.h>
-
-
 /**
 * linear_search - traverse an array to find the index of an element
 *
@@ -10,8 +7,8 @@
 *
 * Returns: the index of the number if found, else -1
 */
-int linear_search(int numbers[], size_t count, int number) {
-    for (size_t i = 0; i < count; ++i) {
+int linear_search(int numbers[], int count, int number) {
+    for (int i = 0; i < count; ++i) {
         if (numbers[i] == number) {
             return i;
         }
